# docextract-ai/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DocExtract AI
  description: Production-grade document intelligence SaaS API for Indian GST documents.
  version: 1.0.0
  contact:
    name: DocExtract AI Support
servers:
  - url: https://api.docextract.ai
  - url: http://localhost:8000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
  /api/v1/auth/token:
    post:
      tags: [auth]
      security: []
      summary: Exchange email + password for a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string, default: bearer }
                  expires_in: { type: integer }
  /api/v1/auth/api-key:
    post:
      tags: [auth]
      summary: Mint a new API key for the current tenant (JWT required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                rate_limit_per_minute: { type: integer, default: 60 }
      responses:
        "201":
          description: API key minted (shown once)
  /api/v1/auth/api-keys:
    get:
      tags: [auth]
      summary: List API keys for the current tenant (JWT required; never returns plaintext)
      parameters:
        - { in: query, name: include_revoked, schema: { type: boolean, default: false } }
      responses: { "200": { description: list of keys } }
  /api/v1/auth/api-keys/{key_id}/rotate:
    post:
      tags: [auth]
      summary: Rotate an API key — old secret stops working immediately
      parameters:
        - { in: path, name: key_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: New plaintext secret (shown once)
        "404": { description: api_key_not_found }
        "409": { description: api_key_revoked }
  /api/v1/auth/api-keys/{key_id}:
    delete:
      tags: [auth]
      summary: Soft-revoke an API key (idempotent)
      parameters:
        - { in: path, name: key_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: revoked }
        "404": { description: api_key_not_found }
...